More precise mapping of scaled face coordinates to camera frame coordinates
Let's consider the following face detection:

   Face Found on layer 2 in 51x38 at X: 17, Y: 3

That coordinates should be translated to

   X0: 17 * 320 / 51,   Y0:  3 * 240 / 38
   W0: 24 * 320 / 51,   H0: 24 * 240 / 38

That is (106.6, 18.9) with size (150.6, 151.6) or

  x0 = 106.6 .. 257.2
  y0 =  18.9 .. 170.5

Take fractions of the pixels found by ceiling window width and height to get:

  x0 = 106 .. 258
  y0 =  18 .. 171

This fixes an effect of loosing some pixels on the right and the bottom edges
of the frame during face detection.  It also gives face sizes closer to the
ones expected by ReID pipeline.

YouTrack: GW-246